More changes.
More changes. The bot now gets an authtoken and connects to dAmn. Event
system is operational, as is command system. Bot has two commands at the
moment (about and ping) and parses a few basic dAmn commands. Tablumps
still need to be parsed. (I'll do it tomorrow or this week)